feat: error boundary, App/Sidebar refactor, drag-drop fix (#14)
- Add ErrorBoundary with fallback UI and ConnectionStatus banner
- Extract modal state to Zustand store (modalStore)
- Extract WebSocket handler to useRealtimeSync hook
- Extract clipboard links, collection variables, Tauri close, cURL panel
- Extract Sidebar drag-and-drop to useSidebarDragDrop hook
- Extract JsonEditor helpers to utils (jsonLinter, jsonComments, jsonEditorTheme)
- Add optimistic UI updates for drag-and-drop reorder/move
- Fix JSON comment preservation on closing bracket lines
- Add E2E tests for error boundary and drag-and-drop operations
- App.jsx: 1270 → 619 lines, Sidebar.jsx: 1275 → 995 lines, JsonEditor.jsx: 473 → 183 lines
