+++ src/stores/modalStore.js
@@ -1,0 +1,27 @@
+import { create } from 'zustand';
+
+const useModalStore = create((set) => ({
+  // Modal visibility
+  showEnvEditor: false,
+  showImportCurl: false,
+  showSettings: false,
+  showAbout: false,
+  showCloseModal: false,
+
+  // Tab close modals
+  draftSavePending: null,
+  tempCloseTabId: null,
+  dirtyCloseTabId: null,
+
+  // Setters
+  setShowEnvEditor: (v) => set({ showEnvEditor: v }),
+  setShowImportCurl: (v) => set({ showImportCurl: v }),
+  setShowSettings: (v) => set({ showSettings: v }),
+  setShowAbout: (v) => set({ showAbout: v }),
+  setShowCloseModal: (v) => set({ showCloseModal: v }),
+  setDraftSavePending: (v) => set({ draftSavePending: v }),
+  setTempCloseTabId: (v) => set({ tempCloseTabId: v }),
+  setDirtyCloseTabId: (v) => set({ dirtyCloseTabId: v }),
+}));
+
+export default useModalStore;
